+++ InterferometricSignalsProcessing/Filters/GradientDescentFilterIS1D.cpp
@@ -27,7 +27,7 @@ void GradientDescentFilterIS1D::setState(Eigen::Vector4d st)
 	state = st;
 }
 
-double GradientDescentFilterIS1D::h(Eigen::Vector4d st)
+float GradientDescentFilterIS1D::h(Eigen::Vector4d st)
 {
 	return st(0) + st(1)*cos(st(3));
 }
@@ -37,7 +37,7 @@ Eigen::Vector4d GradientDescentFilterIS1D::f(Eigen::Vector4d st)
 	return st + Eigen::Vector4d(0, 0, 0, 2 * M_PI*st(2));
 }
 
-Eigen::Vector4d GradientDescentFilterIS1D::gradient(double obs, Eigen::Vector4d st)
+Eigen::Vector4d GradientDescentFilterIS1D::gradient(float obs, Eigen::Vector4d st)
 {
 	if (obs - h(st) > 0)
 		return Eigen::Vector4d(1, cos(st(3)), 0, -st(1)*sin(st(3)));
@@ -45,7 +45,7 @@ Eigen::Vector4d GradientDescentFilterIS1D::gradient(double obs, Eigen::Vector4d 
 		return Eigen::Vector4d(1, cos(st(3)), 0, -st(1)*sin(st(3)))*-1;
 }
 
-void GradientDescentFilterIS1D::estimate(double obs, StopCriterion criterion, bool doPrediction)
+void GradientDescentFilterIS1D::estimate(float obs, StopCriterion criterion, bool doPrediction)
 {
 	if (doPrediction)
 		state = f(state);
@@ -75,12 +75,12 @@ void GradientDescentFilterIS1D::estimate(double obs, StopCriterion criterion, bo
 	}
 }
 
-double GradientDescentFilterIS1D::evaluateSignalValue()
+float GradientDescentFilterIS1D::evaluateSignalValue()
 {
 	return h(state);
 }
 
-int GradientDescentFilterIS1D::sign(double s)
+int GradientDescentFilterIS1D::sign(float s)
 {
 	if (s > 0.00000000001)
 		return 1;
